security_fixed

+++ ecomsboot/src/main/java/org/ecom/ecomsboot/model/Cart.java
@@ -21,8 +21,7 @@ public class Cart {
     @OneToMany(mappedBy = "cart", cascade = CascadeType.ALL, orphanRemoval = true)
     private Set<CartItem> cartItems = new HashSet<>();
 
-    private double totalPrice;
-
+    private int totalPrice;
 
     private int totalItem;
